requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
mysql-connector-python==8.3.0
plotly==5.18.0
openpyxl==3.1.2
python-dotenv==1.0.1
SQLAlchemy==2.0.27
pymysql==1.1.0 